+++ apps/meteor/app/livechat/server/methods/setUpConnection.ts
@@ -3,7 +3,7 @@ import type { ServerMethods } from '@rocket.chat/ddp-client';
 import { check } from 'meteor/check';
 import { Meteor } from 'meteor/meteor';
 
-import { Livechat } from '../lib/LivechatTyped';
+import { notifyGuestStatusChanged } from '../lib/guests';
 
 declare module '@rocket.chat/ddp-client' {
 	// eslint-disable-next-line @typescript-eslint/naming-convention
@@ -33,7 +33,7 @@ Meteor.methods<ServerMethods>({
 		if (this.connection && !this.connection.livechatToken) {
 			this.connection.livechatToken = token;
 			this.connection.onClose(async () => {
-				await Livechat.notifyGuestStatusChanged(token, UserStatus.OFFLINE);
+				await notifyGuestStatusChanged(token, UserStatus.OFFLINE);
 			});
 		}
 	},
